+++ source/Private/Task.Generate_External_Help_File_For_Public_Commands.ps1
@@ -1,0 +1,15 @@
+<#
+    .SYNOPSIS
+        This is the alias to the build task Generate_External_Help_File_For_Public_Commands's script file.
+
+    .DESCRIPTION
+        This makes available the alias 'Task.Generate_External_Help_File_For_Public_Commands' that is
+        exported in the module manifest so that the build task can be correctly
+        imported using for example Invoke-Build.
+
+    .NOTES
+        This is using the pattern lined out in the Invoke-Build repository
+        https://github.com/nightroman/Invoke-Build/tree/master/Tasks/Import.
+#>
+
+Set-Alias -Name 'Task.Generate_External_Help_File_For_Public_Commands.build' -Value "$PSScriptRoot/tasks/Generate_External_Help_File_For_Public_Commands.build.ps1"